Moved all methods out of deprecated XLIFF
