+++ Tests.GoogleVertexAI/EditActionTests.cs
@@ -1,6 +1,9 @@
 ﻿using Apps.GoogleVertexAI.Actions;
 using Apps.GoogleVertexAI.Models.Requests;
+using Apps.GoogleVertexAI.Polling;
+using Apps.GoogleVertexAI.Polling.Model;
 using Blackbird.Applications.Sdk.Common.Files;
+using Blackbird.Applications.Sdk.Common.Polling;
 using Blackbird.Applications.SDK.Extensions.FileManagement.Interfaces;
 using GoogleVertexAI.Base;
 using Newtonsoft.Json;
@@ -10,7 +13,7 @@ namespace Tests.GoogleVertexAI;
 [TestClass]
 public class EditActionTests : TestBase
 {
-    private const string ModelName = "gemini-2.0-flash";
+    private const string ModelName = "gemini-2.5-flash";
 
 
     [TestMethod]
@@ -52,5 +55,55 @@ public class EditActionTests : TestBase
         Assert.IsNotNull(result.EditedText);
         Console.WriteLine("Original: " + localizeRequest.TargetText);
         Console.WriteLine("Localized: " + result.EditedText);
+    }
+
+    [TestMethod]
+    public async Task Batch_edit_text_returns_valid_xliff()
+    {
+        var actions = new EditActions(InvocationContext, FileManager);
+        var file = new FileReference { Name = "contentful.html.xlf" };
+        var translateRequest = new EditFileRequest
+        {
+            File = file,
+            AIModel = ModelName,
+        };
+        string? systemMessage = null;
+        var glossaryRequest = new GlossaryRequest();
+
+        var startbatchResopnse = await actions.BatchEditContent(translateRequest, new PromptRequest { }, systemMessage, glossaryRequest);
+        Assert.IsNotNull(startbatchResopnse);
+        Console.WriteLine(startbatchResopnse.JobName);
+
+        var polling = new BatchPolling(InvocationContext);
+
+
+        var result = await polling.OnBatchFinished(new PollingEventRequest<BatchMemory>() {
+            Memory = new BatchMemory
+            {
+                LastPollingTime = DateTime.UtcNow,
+                Triggered = false
+            }
+        }, new BatchIdentifier { JobName = startbatchResopnse.JobName });
+
+        while (!result.FlyBird)
+        {
+            await Task.Delay(3000);
+            result = await polling.OnBatchFinished(new PollingEventRequest<BatchMemory>()
+            {
+                Memory = new BatchMemory
+                {
+                    LastPollingTime = DateTime.UtcNow,
+                    Triggered = false
+                }
+            }, new BatchIdentifier { JobName = startbatchResopnse.JobName });
+        }
+
+        var batchActions = new BatchActions(InvocationContext, FileManager);
+
+        var finalResult = await batchActions.DownloadXliffFromBatch(startbatchResopnse.JobName, new GetBatchResultRequest { OriginalXliff = startbatchResopnse.TransformationFile });
+
+        Console.WriteLine(JsonConvert.SerializeObject(finalResult, Formatting.Indented));
+
+        Assert.IsNotNull(finalResult.File);
     }
 }
